# Optimized addition, subtraction and negation