--- a/src/multiply.jl
+++ b/src/multiply.jl
@@ -1,7 +1,7 @@
-@inline _alloc_temp_array(size_1::Int, x::AbstractVector{T}) where T = zeros(T, size_1)
-@inline _alloc_temp_array(size_1::Int, x::AbstractMatrix{T}) where T = zeros(T, size_1, size(x, 2))
+@inline _alloc_temp_array(size_1::Int, x::AbstractVector{T}) where {T} = zeros(T, size_1)
+@inline _alloc_temp_array(size_1::Int, x::AbstractMatrix{T}) where {T} = zeros(T, size_1, size(x, 2))
 
-const MatrixOrFactorization{T} = Union{AbstractMatrix{T}, Factorization{T}}
+const MatrixOrFactorization{T} = Union{AbstractMatrix{T},Factorization{T}}
 
 for N in (1, 2)
     for op in [:mul, :ldiv]
@@ -14,17 +14,17 @@ for N in (1, 2)
             op_expr_square = :(@views q[slc, :] = $op!(temp[1:n, :], m, q[slc, :]))
         end
 
-        @eval function $square_kernel!(temp::AbstractArray{T1, $N}, q::AbstractArray{T2, $N}, n::Int, i_left::Int, m::MatrixOrFactorization{T3}, i_right::Int) where {T1,T2,T3}
+        @eval function $square_kernel!(temp::AbstractArray{T1,$N}, q::AbstractArray{T2,$N}, n::Int, i_left::Int, m::MatrixOrFactorization{T3}, i_right::Int) where {T1,T2,T3}
             # apply kron(I(l), m, I(r)) where m is square to the given vector x, overwriting x in the process
 
-            if m isa Factorization || m isa Adjoint{T3, <:Factorization} || m isa Transpose{T3, <:Factorization} || m != I
+            if m isa Factorization || m isa Adjoint{T3,<:Factorization} || m isa Transpose{T3,<:Factorization} || m != I
                 # if matrix is the identity, skip matmul/div, unless it's some sort factorization, then proceed anyway
                 irc = i_right * n
 
                 base_i, top_i = 0, (irc - i_right)
                 @inbounds for i_l in 1:i_left
                     for i_r in 1:i_right
-                        slc = base_i + i_r : i_right : top_i + i_r
+                        slc = base_i+i_r:i_right:top_i+i_r
 
                         $op_expr_square
                     end
@@ -37,7 +37,7 @@ for N in (1, 2)
         end
 
         square_func! = Symbol("_kron_", op, "_fast_square!")
-        @eval function $square_func!(out::AbstractArray{T1, $N}, x::AbstractArray{T2, $N}, matrices) where {T1,T2}
+        @eval function $square_func!(out::AbstractArray{T1,$N}, x::AbstractArray{T2,$N}, matrices) where {T1,T2}
             ns::Vector{Int} = [size(m, 1) for m in matrices]
             i_left::Int = 1
             i_right::Int = prod(ns)
@@ -71,7 +71,7 @@ for N in (1, 2)
         end
 
         rect_kernel! = Symbol("_kron_", op, "_kernel_rect!")
-        @eval function $rect_kernel!(temp::AbstractArray{T1, $N}, q::AbstractArray{T2, $N}, r_h::Int, c_h::Int, i_left::Int, m::MatrixOrFactorization{T3}, i_right::Int) where {T1,T2,T3}
+        @eval function $rect_kernel!(temp::AbstractArray{T1,$N}, q::AbstractArray{T2,$N}, r_h::Int, c_h::Int, i_left::Int, m::MatrixOrFactorization{T3}, i_right::Int) where {T1,T2,T3}
             # apply kron(I(i_left), m, I(i_right)) to the given vector q
 
             # don't bother checking for identity, since we know the matrix
@@ -86,8 +86,8 @@ for N in (1, 2)
             top_i, top_j = (irc - i_right), (irr - i_right)
             @inbounds for i_l in 1:i_left
                 for i_r in 1:i_right
-                    slc_in  = base_i + i_r : i_right : i_r + top_i
-                    slc_out = base_j + i_r : i_right : i_r + top_j
+                    slc_in = base_i+i_r:i_right:i_r+top_i
+                    slc_out = base_j+i_r:i_right:i_r+top_j
 
                     $op_expr_rect
                 end
@@ -105,7 +105,7 @@ for N in (1, 2)
         rect_func! = Symbol("_kron_", op, "_fast_rect!")
         ri = (op == :mul) ? 1 : 2
         ci = (op == :mul) ? 2 : 1
-        @eval function $rect_func!(out::AbstractArray{T1, $N}, x::AbstractArray{T2, $N}, matrices) where {T1,T2}
+        @eval function $rect_func!(out::AbstractArray{T1,$N}, x::AbstractArray{T2,$N}, matrices) where {T1,T2}
             r::Vector{Int} = [size(m, $ri) for m in matrices]
             c::Vector{Int} = [size(m, $ci) for m in matrices]
             i_left::Int = 1
@@ -130,7 +130,7 @@ for N in (1, 2)
         end
     end
 
-    @eval function _kronsum_mul_fast!(out::AbstractArray{T1, $N}, x::AbstractArray{T2, $N}, matrices) where {T1,T2}
+    @eval function _kronsum_mul_fast!(out::AbstractArray{T1,$N}, x::AbstractArray{T2,$N}, matrices) where {T1,T2}
         ns::Vector{Int} = [size(m, 1) for m in matrices]
         i_left::Int = 1
         i_right::Int = prod(ns)