This PR reduces some temporary allocations in addition and subtraction, and optimizes unary negation for `KroneckerProduct`.

Master:
```julia
julia> A = rand(30,30) ⊗ rand(40, 40);

julia> B = rand(40,40) ⊗ rand(30,30);

julia> C = rand(40,40) ⊗ Diagonal(rand(30));

julia> @btime $A + $B;
  5.653 ms (6 allocations: 32.96 MiB)

julia> @btime $A - $B;
  5.836 ms (6 allocations: 32.96 MiB)

julia> @btime -$A;
  2.055 ms (2 allocations: 10.99 MiB)

julia> @btime -$C;
  3.896 ms (2 allocations: 10.99 MiB)
```

This PR:
```julia
julia> @btime $A + $B;
  4.846 ms (4 allocations: 21.97 MiB)

julia> @btime $A - $B;
  4.920 ms (4 allocations: 21.97 MiB)

julia> @btime -$A; # lazy kronecker product
  1.086 μs (1 allocation: 7.19 KiB)

julia> @btime -$C; # lazy kronecker product
  84.321 ns (1 allocation: 304 bytes)
```

Sorry about the whitespace changes, it's an automated VSCode setting that I've not figured out how to disable. The main changes are lines 411-494 in src/base.jl, and an added `lastmatrix` method in kroneckerpowers.jl on line 44