@MichielStock 2022-01-29
Looks good and the speedups are even more impressive on my computer.

I don't think the whitespace changes are too bad. If you want, you might even open the other files of the source code so they are consistent?